Create the complete recipe website
To complete the project, I have uploaded 3 images and 4 html files. I
put the images in the img folder and the recipe html files in the
recipes folder for clarity. I have a homepage that links to all 3
recipe pages (html files). Each recipe page has the name of the recipe,
an image of the finished product, a description of the recipe, an
ingredients list, and the actual steps of the recipe.

All of this content was made for a proper project submission. Please
feel free to have a look. This project is my first big step on the Odin
Project and web development. Hopefully, I'll raise a toast to many more.
:)